docs/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
pydantic
